requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
pyjwt
jinja2